+++ Source/dev/checksum.c
@@ -2,7 +2,7 @@
 
 WORD Checksum16(BYTE *apbDst, UINT anNum)
 {
-	CHKSUM_ADR = (UINT)apbDst;
+	CHKSUM_ADR = (intptr_t)apbDst;
 	CHKSUM_LEN = anNum;
 	CHKSUM_GO = 1;
 	while (CHKSUM_GO);
